pyproject.toml
[project]
name = "docker-postgres-backup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic-settings>=2.6.1",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.3",
    "rich>=13.9.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "python-unittest>=0.2.12",
]

[project.scripts]
dpb = "docker_postgres_backup.main:main"
